moved API to single function, added new async functions that behave differently to hooks - can be used with forms and submissions
