+++ frontend/src/App.jsx
@@ -6,8 +6,8 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 // import the files for each URL endpoint
 import SellerSignupPage from "./authorisationPages/sellerSignupPage.jsx";
 import BuyerSignupPage from "./authorisationPages/buyerSignupPage.jsx";
-import LoginPage from "./authorisationPages/loginPage.jsx";
-import GamePage from "./gamePages/game";
+import SellerLoginPage from "./authorisationPages/sellerLoginPage.jsx";
+import GamePage from "./gamePages/game.jsx";
 import MarketplacePage from "./marketplacePages/marketplace";
 import PageNotFound from "./reusableComponents/pageNotFound";
 import UserHomePage from "./homePage/userHomePage.jsx";
@@ -22,8 +22,8 @@ function App() {
     <Router>
       <Routes>
         <Route path="/user" element={<UserHomePage />} />
-        <Route path="/" element={<LoginPage />} />
-        <Route path="/login" element={<LoginPage />} />
+        {/* <Route path="/" element={<LoginPage />} /> */}
+        <Route path="/seller/login" element={<SellerLoginPage />} />
         <Route path="/signup/buyer" element={<BuyerSignupPage />} />
         <Route path="/signup/seller" element={<SellerSignupPage />} />
         <Route path="/user" element={<UserHomePage />} />
